perf(utils): Use Set to make arrayDiff O(n+m) vs O(n*m)
This is used in a couple places where the values for n and m could be medium size (e.g. when
checking for new partitions within a topic), and we've seen it appear in some of our CPU profiling.
So making this linear instead of quadratic should help slightly when checking topics with higher
partition counts.